Version 4.8
- Added: New advanced video thumbnail generation which is vastly faster
and more stable than the previous feature. No more unexpected crashes
due to bxsdk32.dll or ffmpeg.exe on some machines. We developed a
library to directly access video files so there will be no more side
effects of launching an external command line exe from IIS. The setting
"FileUltimate:DisableVideoThumbnails" in web.config still works but as
we solved the stability problems, you should not need to disable
thumbnails for video files anymore. We may offer this high-performance
and stable video frame reading library as a separate product soon.

- Improved: The picture quality of the generated video thumbnails, you
will notice that the duration text on the bottom right is now more crisp
and readable.