Version 5.1.5
- Added: "Print" permission for controlling print, download as pdf and
text selection actions in the document viewer.
- Changed: Removed DisablePreviewerPrinting and
DisablePreviewerTextSelection properties as these will now be controlled
by the new Print permission per folder-basis.
- Improved: The Document Viewer;
- Added: "Download" and "Download as PDF" buttons to the toolbar.
"Download" button will be disabled when there is no Download permission
and "Download as PDF" button
will be disabled when there is no Print permission. This is because,
downloading a DOCX file as a PDF file is
considered same as printing that document.
- Improved: The toolbar is now reponsive, buttons will not be hidden and
they will flow naturally to next row below.
So all the available buttons will be visible regardless of the viewer
size. Also the buttons are spaced better.
- Fixed: Some scanned documents were displayed as blurred, i.e. the text
was unreadable.
- Fixed: Some XPS files were not being displayed at all.
- Fixed: Even when the printing was disabled, the user could still open
"Print Dialog" by pressing CTRL + P.
- Added: New examples engine with flat/modern theme.