Version 5.3.8
Version 5.3.8 - May 22, 2016
- Added: ImageViewerEnabled, MediaPlayerEnabled, DocumentViewerEnabled,
ArchiveBrowsingEnabled properties
so that you can disable the corresponding features if required. Also
moved ImageThumbnailsEnabled and
VideoThumbnailsEnabled properties from FileUltimateConfiguration to
FileManager class so that they can be disabled
on instance-level and not for whole application.
- Improved: When access is denied to the cache path, a more
user-friendly error message with instructions will be displayed.
Also the physical path will not be included in the message for security
purpose.
- Improved: Updated GleamTech.Core for allowing access via IP for domain
restricted licenses.
- Fixed: In Web Site projects, when you added FileManager markup in an
aspx page and built the web site, the build failed
with the following error message (actually 3 error messages but main one
is this, others are consequences):
Cannot create an object of type 'GleamTech.FileSystems.Location' from
its string representation.
- Fixed: When FileManager was inserted in a page with AngularJS,
clicking on context menu items did not work and instead
caused a redirect to the home page. This was because AngularJS took over
all <a href="#"> links on the page and as
AngularJS considered them as a navigation and as there's no route
specified it redirected to the home page. To be
compatible with AngularJS, from now on menu items will use <a
href="javascript:"> as links.
- Fixed: When in details view, clicking on "Type" column did not sort
the files.
- Fixed: Unnecessary "Path cannot be discovered" error when a root
folder had no permissions.

Version 5.3.6 - May 16, 2016
- Fixed: Moving a folder which had files at second and higher level
depth failed.
- Fixed: FileManagerRootFolder.GetFolder was broken when the root folder
had no access control.
- Added: TotalSize, ElapsedTime and TransferRate properties to
FileManagerDownloadedEventArgs so that more information
will be available in Downloaded event.
- Improved: Updated media player, document conversion, graphics and
video engines.

Version 5.3.5 - May 11, 2016
- Fixed: Some problems continued on Internet Explorer and Microsoft Edge
browsers after the new drag and drop feature.
For example, columns on the file list were not clickable anymore. With
this version all problems related to clicking
should be fixed and clicking should feel more responsive again.
- Fixed: Some UI quirks. For example error window was still displaying
behind the background mask on some cases and
minimize/restore state of a window was not stable (scrolling the main
view behind the window caused it the window
to minimize).
- Improved: Made the default action of dragged items "Move to" even if
the drag is made across view (eg. dragging from
main view to folder tree). So "Copy to" action will be switched on only
if CTRL key is pressed. This exactly mimics
"Windows Explorer" behaviour.
- Fixed: After you drag and drop in the folder tree, you could get an
error message when you clicked the parent folder.
When dragging, the tree was confused and changed the current selection
to the dragged one (focused) although this was
not visually reflected so we had a ghost selection.
- Fixed: Prevented horizontal scroll of folder tree when dragging on the
right edge and waiting for 500ms. The scroll helper
should only work vertically for the folder tree.
- Fixed: While dragging on the bottom of the main view or folder tree,
the page would flicker.
- Improved: Updated PDF icon with the most recent one. Also fixed large
icons for ascx, ashx, asmx, aspx, config, css, htm,
js, resx, xsd, xsl, xml, h extensions so they will not be pixelated (not
upsized from small icon).
- Fixed: Icons for 1 letter extensions were not shown (eg. .c and .h),
also "Type" column showed "File" instead of e.g.
"C File".
- Added: Wildcard support for FileManagerAccessControl.Path property.
Now you can use these wildcards in a path:
The ? wildcard character matches a single character.
The * wildcard character matches zero or more characters.
The ** wildcard character sequence matches a partial path.
For example, if you set Path="\**\*MatchMe*", this access control will
appy to the folders containing MatchMe keyword
in name under both root folder and under root folder's all subfolders
(recursive).