Version 5.5.5
## Version 5.5.5 - October 17, 2016

- **Added:** Visual Studio Design View support. From now on, an image
preview (placeholder) of FileManager control
will be displayed in Design View. This preview is also resizable and it
directly reflects Width and Height properties.

- **Fixed:** Visual Studio Design View error:
"Cannot create an object of type 'GleamTech.FileSystems.Location' from
its string representation..."

- **Fixed:** Amazon S3 root folders can now list more than 1000 items.
Also improved listing performance of folders so
you should not have any problems with huge buckets.

- **Fixed:** When Amazon S3 bucket was empty, it caused an error.

- **Fixed:** ClientLoading event was being fired too late, some
functions like fileManager.onViewRefresh are registered
as event handlers to internal components so even you override them later
the components still referred to the old
function reference, so Ext.override was not working. From now on
ClientLoading will now fire at the start of
initComponent and ClientLoaded at the end of initComponent.
ClientLoading should be used for Ext.override calls,
ClientLoaded should be called for adding actions, UI elements.

- **Fixed:** IE11 on touch laptops were detected as a mobile device so
Document Viewer was displayed in mobile mode.
Improved mobile device detection to prevent this kind of issues.

- **Improved:** Document Viewer. Some formats (e.g. common image formats
and digital paper formats like XPS and EPUB)
will be opened faster for first-time view. More formats can be viewed
with Document Viewer.

- **Improved:** Updated Media Player. Also updated the skin to youtube
style.

- **Improved:** Image and document caches will now group cached files
for the same inputs in subfolders.
Also the subfolder name  will contain the original file name so it will
be easier to locate the corresponding cached
files for an input. The old cached files will be arranged into the new
structure automatically one by one when they
are accessed so no action is required for the developer.