Version 5.9.2
## Version 5.9.2 - May 2, 2017

- **Fixed:** Clicking on UI items was not working on Firefox 52+ when
used on a touch-enabled device.
Firefox changed a default browser behavior/setting in recent versions
which broke clicking behavior
in some products: If you have a touch enabled laptop or PC monitor and
when you use a mouse,
your clicks were not received by the controls displayed in Firefox. The
problem did not happen if you
did not use a touch-enabled device.

- **Improved:** Document Viewer:

- **Fixed:** Sometimes DocumentViewer's vertical scrollbar's bottom
button was not visible.

- **Fixed:** Especially on IE, DocumentViewer's pan tool (hand cursor)
was overlapping scrollbar so it made the user
think the hotspot was on the scrollbar but when the user moved the mouse
the user was actually still panning the
document because the hotspot of the hand cursor was still on the
document and not on the scrollbar.
This confused the user thinking into clicking on scrollbar caused
"jumping back to the first page of the document".
Now, when the user approaches a scrollbar, the hand cursor will be
changed to default cursor and panning too near
the scrollbar will be disabled. This way panning and scrolling actions
will never be confused.

- **Fixed:** When printing on IE, an extra blank page was being added at
the end.

- **Improved:** Viewing accuracy for Presentation and Imaging formats.