Version 5.10.0
## Version 5.10.0 - June 20, 2017

- **Added:** New "Item check boxes" mode for convenient multiple
selection of files/folder especially on touch devices. The option can be
toggled on "View" tab of ribbon. For tablet and phone devices this
option will be turned on by default because drag-selecting is hard (long
press and scroll) on these devices (unlike desktop where mouse is
available). So check-selecting is better for these devices.
Check-selecting works with all (6) view layouts.

- **Fixed:** Thumbnails were not loading on iOS (retina).