Version 5.18.5
## Version 5.18.5 - February 12, 2018

- **Fixed:** TIFF file support in DocumentViewer.
- Since v5.14.0, the rendering speed of TIFF files was slower and the
created .xpz and .pdf files were bigger.
This was due to the fix "Viewing some TIFF files caused crashing of IIS
worker process". Now the performance
is restored, file sizes are optimized and crash is fixed at the same
time.

- TIFF files with an EXIF thumbnail (e.g. saved with Photoshop) caused
"Image loading failed" error or
or before v5.14.0, the thumbnail stored in the EXIF was shown but not
the high res TIFF file.
Now the original TIFF image will be displayed as expected.

- Some rare TIFF formats like WANG TIFF (Wang JPEG images encapsulated
in TIFF files by Microsoft Office imaging Version)
were not recognized.

- If the file extension was .tiff/.tif but the actual format was another
image format like PNG, the file was not loaded.
Now even if the file is wrong image format, it will be loaded and
displayed.