Version 5.19.0
## Version 5.19.0 - February 17, 2018

- **Improved:**  MVC partial view support. Old component instance with
the same name will be destroyed so that
the new component is rendered successfully. Added MVC partial view
example to MVC example projects.

- **Added:** Clicking icon in the breadcrumb navigation bar (going to
Home) will now list the root folders instead
of showing blank view which was confusing to some users. So now root
folders will be visible in the right view
just like folders.

- **Added:** InitialFolder property will now expand all subfolders in
the path not only the final subfolder.

- **Improved:** Rendering speed for "Large icons" and "Medium icons"
view layout.
For cross-browser multi-line clamping, JS was used but now switched to
pure CSS instead because
clamping already rendered text from JS could stall IE for a second or
so, although it was not noticable in Chrome.
So now, for overflowing item names, ellipsis will be shown in Chrome as
it's the only browser that supports it
natively but for other browsers only consistent multi-line size will be
maintained without ellipsis
(which is worth for better performance).

- **Fixed:** Maximizing, resizing and then closing a child window (e.g.
DocumentViewer window) left a mask on
top of FileManager.