Version 5.21.5
## Version 5.21.5 - April 7, 2018

- **Fixed:** In some cases, ASP.NET's UrlRoutingModule was still taking
over the handlers (HTTP 404) after
the new own routing was added in v5.21.0. This is now completely fixed.
For reference, it was avoidable with v5.21.0 via adding the below line
at the start of RegisterRoutes method
in RouteConfig.cs (of MVC project):

routes.IgnoreRoute("{resource}.ashx/{*pathInfo}")

Now, this is not required anymore starting with this version.

- **Improved:** Document Viewer. Newer versions of Cad formats (AutoCad
2018) is now supported.