Version 6.0.0
## Version 6.0.0 - May 22, 2018

- **Added:** "ASP.NET Core on .NET Framework" support. ASP.NET Core uses
a completely new web engine different than
System.Web but now you can use GleamTech products on both classic
ASP.NET and ASP.NET Core with the same net40 DLL
as we abstracted classes like HttpContext, HttpRequest and HttpResponse
etc. and we implemented a Middleware for
mimicking HttpModule and HttpHandler. So our DLL auto-magically works
regardless of whether you are running
under classic ASP.NET or ASP.NET Core. Note that "ASP.NET Core on .NET
Core" is a different platform and it's not
supported yet because it requires porting all the code from .NET
Framework runtime to .NET Core runtime so it
requires a new DLL and .NET Core runtime does not provide all of the
APIs yet. Minimum supported version is
ASP.NET Core MVC 2.0.3 on .NET Framework 4.6.1 (this is because 2.0.3
fixes a bug related to referencing
external DLLs in a razor page).

- **Added:** New example project for "ASP.NET Core on .NET Framework",
please refer to it for info on sample usage.

- **Added:** Documentation for the library (similar to
DocumentUltimate). API docs is not complete yet but important
classes are covered.

- **Changed:** GleamTech.FileUltimate namespace is mainly moved to
GleamTech.FileUltimate.AspNet.UI

- GleamTech.FileUltimate.AspNet.UI: Sub-namespace for common FileManager
component model.
- GleamTech.FileUltimate.AspNet.WebForms: Sub-namespace for ASP.NET Web
Forms support.

This mimicks DocumentUltimate namespaces.

- **Changed:** License keys are changed so please go to
https://www.gleamtech.com/upgrade and acquire a new license
key if you want to use this version (or higher). If your one year
maintenance has not ended, you will receive a
new free license key on the same page.

- **Improved:** Updated Media Player.

- **Improved**: Document Viewer stability, accuracy for Portable,
Presentation and Spreadsheet formats.